perf: 优化构建脚本

+++ build.sh
@@ -4,10 +4,12 @@ DOCKER_IMAGE_PREFIX="swr.cn-north-4.myhuaweicloud.com/nineaiyu"
 
 images="node:22.10.0-alpine xadmin-node:22.10.0-alpine"
 for image in ${images};do
-  full_image_path="${DOCKER_IMAGE_PREFIX}/${image}"
-  docker pull  "${full_image_path}"
-  docker tag "${full_image_path}" "${image}"
-  docker rmi -f "${full_image_path}"
+  if ! docker images --format "{{.Repository}}:{{.Tag}}" |grep "${image}" &>/dev/null;then
+    full_image_path="${DOCKER_IMAGE_PREFIX}/${image}"
+    docker pull  "${full_image_path}"
+    docker tag "${full_image_path}" "${image}"
+    docker rmi -f "${full_image_path}"
+  fi
 done
 
 # clean old build docker and image
